Improve RO disk UT
